Refactor commands and update aliases
